fix(common): Consolidate Compat Impls Into Transaction Module (#2208)
* fix(common): Consolidate Compat Impls Into Transaction Module

Split `abstraction.rs` into `core.rs` and `traits.rs`, moved all compat
impls from the top-level `compat.rs` and `consensus_compat.rs` orphan
modules directly into `transaction/core.rs` where they belong alongside
the `OpTransaction` type, and fixed the stale `super::abstraction` import
in `builder.rs` that would cause a compile failure.

* fix(common): Fix nightly fmt import ordering
